requirements.txt
# ── Web Framework ────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.1.0

# ── Environment ───────────────────────────────────────────────
python-dotenv==1.0.1

# ── AI / Gemini ───────────────────────────────────────────────
google-generativeai==0.4.1

# ── PDF Extraction ────────────────────────────────────────────
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0

# ── Excel / CSV ───────────────────────────────────────────────
openpyxl==3.1.2

# ── Report Generation ─────────────────────────────────────────
reportlab==4.1.0

# ── Web Search (Research Agent — no API key needed) ───────────
duckduckgo-search==4.4.3
httpx==0.26.0

# ── NLP / Text ────────────────────────────────────────────────
spacy==3.7.2

# ── Utilities ─────────────────────────────────────────────────
aiofiles==23.2.1